requirements.txt
flask==2.3.3
langchain>=0.1.0
langchain-google-genai>=0.0.5
google-generativeai>=0.3.2
langchain-community>=0.0.10
langchain-core>=0.1.10
requests>=2.31.0
python-dotenv>=1.0.0
ddgs>=8.0.0
google-api-python-client>=2.0.0
PyGithub>=2.0.0 